refactor: prereq
